requirements.txt
absl-py==0.8.1
astor==0.8.0
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.2
opt-einsum==3.1.0
Pillow==6.1.0
protobuf==3.10.0
pyparsing==2.4.2
python-dateutil==2.8.0
scipy==1.3.1
six==1.12.0
tensorboard==2.0.0
tensorboardX==1.9
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
termcolor==1.1.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.36.1
Werkzeug==0.16.0
wrapt==1.11.2
